refactor(toast): update file structure
