--- packages/admin-ui/src/components/Toast/index.tsx
@@ -1,2 +1,0 @@
-export { ToastProvider, useToast } from './components/ToastContext'
-export type { Toast as ToastParams } from './types'
