+++ packages/admin-ui/src/index.ts
@@ -31,6 +31,7 @@ export * from './grid'
 export * from './columns'
 export * from './card'
 export * from './visually-hidden'
+export * from './toast'
 
 export * from './components/Skeleton'
 export * from './components/Label'
@@ -47,7 +48,6 @@ export * from './components/Modal'
 export * from './components/Dropdown'
 export * from './components/Spinner'
 export * from './components/Sidebar'
-export * from './components/Toast'
 export * from './components/Topbar'
 export * from './components/Toolbar'
 export * from './components/Pagination'
